LLm report applied and main improved
LLM report is now in the repository, but I haven't filled it in yet, will do this later. Changes to main are:
relocated my activation functions into my MLP class. Added some other functions into my class like backpropagation and derivative, which still have to be constructed later on. Created a seperate class Architecture, which enables me to easily change the architecture of the model, and allows me to automatically apply all the steps for optimization without having to hard code everything for each hidden layer.